feat: add support for slotted content to dynamic components in `@lwc/ssr-compiler` (#4898)
