+++ packages/@lwc/engine-server/src/__tests__/fixtures/dynamic-components/scoped-styles/expected.html
@@ -1,0 +1,15 @@
+<x-dynamic-component>
+  <template shadowrootmode="open">
+    <x-scoped-style class="lwc-2d8pgjj85pu-host" data-lwc-host-scope-token="lwc-2d8pgjj85pu-host">
+      <template shadowrootmode="open">
+        <style class="lwc-2d8pgjj85pu" type="text/css">
+          div.lwc-2d8pgjj85pu {background-color: wheat;}
+        </style>
+        <div class="foo lwc-2d8pgjj85pu">
+        </div>
+        <div class="bar lwc-2d8pgjj85pu">
+        </div>
+      </template>
+    </x-scoped-style>
+  </template>
+</x-dynamic-component>
