+++ packages/@lwc/engine-server/src/__tests__/fixtures/dynamic-components/scoped-styles/index.js
@@ -1,0 +1,3 @@
+export const tagName = 'x-dynamic-component';
+export { default } from 'x/dynamic';
+export * from 'x/dynamic';
