+++ packages/@lwc/engine-server/src/__tests__/fixtures/dynamic-components/scoped-styles/modules/x/dynamic/dynamic.html
@@ -1,0 +1,3 @@
+<template>
+    <lwc:component lwc:is={ctor}></lwc:component>
+</template>
