+++ packages/@lwc/engine-server/src/__tests__/fixtures/dynamic-components/scoped-styles/modules/x/dynamic/dynamic.js
@@ -1,0 +1,6 @@
+import { LightningElement, track } from 'lwc';
+import ScopedStyleComponent from 'x/scopedStyle';
+
+export default class extends LightningElement {
+    @track ctor = ScopedStyleComponent;
+}
