+++ packages/@lwc/engine-server/src/__tests__/fixtures/dynamic-components/scoped-styles/modules/x/scopedStyle/scopedStyle.html
@@ -1,0 +1,4 @@
+<template>
+    <div class="foo"></div>
+    <div class={clazz}></div>
+</template>
