+++ packages/@lwc/engine-server/src/__tests__/fixtures/dynamic-components/scoped-styles/modules/x/scopedStyle/scopedStyle.js
@@ -1,0 +1,5 @@
+import { LightningElement } from 'lwc';
+
+export default class extends LightningElement {
+    clazz = 'bar'
+}
