+++ packages/@lwc/engine-server/src/__tests__/fixtures/dynamic-components/scoped-styles/modules/x/scopedStyle/scopedStyle.scoped.css
@@ -1,0 +1,3 @@
+div {
+    background-color: wheat;
+}
