+++ packages/@lwc/ssr-compiler/src/__tests__/utils/expected-failures.ts
@@ -28,8 +28,6 @@ export const expectedFailures = new Set([
     'attribute-style/basic/index.js',
     'attribute-style/dynamic/index.js',
     'comments-text-preserve-off/index.js',
-    'dynamic-components/slots/shadow/index.js',
-    'dynamic-components/slots/light/index.js',
     'dynamic-slots/index.js',
     'empty-text-with-comments-non-static-optimized/index.js',
     'if-conditional-slot-content/index.js',
