+++ packages/@lwc/ssr-compiler/src/compile-template/transformers/component/slotted-content.ts
@@ -7,21 +7,21 @@
 
 import { produce } from 'immer';
 import { builders as b, is } from 'estree-toolkit';
-import { kebabcaseToCamelcase, toPropertyName } from '@lwc/template-compiler';
-import { bAttributeValue, getChildAttrsOrProps, optimizeAdjacentYieldStmts } from '../shared';
-import { esTemplate, esTemplateWithYield } from '../../estemplate';
-import { irChildrenToEs, irToEs } from '../ir-to-es';
-import { isNullableOf } from '../../estree/validators';
+import { bAttributeValue, optimizeAdjacentYieldStmts } from '../../shared';
+import { esTemplate, esTemplateWithYield } from '../../../estemplate';
+import { irChildrenToEs, irToEs } from '../../ir-to-es';
+import { isNullableOf } from '../../../estree/validators';
 import type { CallExpression as EsCallExpression, Expression as EsExpression } from 'estree';
 
-import type { BlockStatement as EsBlockStatement } from 'estree';
-import type { Component as IrComponent, ScopedSlotFragment } from '@lwc/template-compiler';
-import type { Transformer } from '../types';
+import type { Statement as EsStatement } from 'estree';
+import type {
+    Component as IrComponent,
+    LwcComponent as IrLwcComponent,
+    ScopedSlotFragment,
+} from '@lwc/template-compiler';
+import type { TransformerContext } from '../../types';
 
-const bYieldFromChildGenerator = esTemplateWithYield`
-    {
-        const childProps = __getReadOnlyProxy(${/* child props */ is.objectExpression});
-        const childAttrs = ${/* child attrs */ is.objectExpression};
+const bGenerateSlottedContent = esTemplateWithYield`
         const slottedContent = {
             light: Object.create(null),
 
@@ -42,21 +42,9 @@ const bYieldFromChildGenerator = esTemplateWithYield`
             }
         }
 
-        ${/* light DOM addContent statements */ is.callExpression}
-        ${/* scoped slot addContent statements */ is.callExpression}
-
-        const scopeToken = hasScopedStylesheets ? stylesheetScopeToken : undefined;
-
-        yield* ${/* generateMarkup */ is.identifier}(
-            ${/* tag name */ is.literal}, 
-            childProps, 
-            childAttrs, 
-            slottedContent,
-            instance,
-            scopeToken,
-        );
-    }
-`<EsBlockStatement>;
+        ${/* light DOM addContent statements */ is.expressionStatement}
+        ${/* scoped slot addContent statements */ is.expressionStatement}
+`<EsStatement[]>;
 
 // Note that this function name (`generateSlottedContent`) does not need to be scoped even though
 // it may be repeated multiple times in the same scope, because it's a function _expression_ rather
@@ -70,14 +58,10 @@ const bAddContent = esTemplate`
     });
 `<EsCallExpression>;
 
-export const Component: Transformer<IrComponent> = function Component(node, cxt) {
-    // Import the custom component's generateMarkup export.
-    const childGeneratorLocalName = `generateMarkup_${toPropertyName(node.name)}`;
-    const importPath = kebabcaseToCamelcase(node.name);
-    cxt.import({ generateMarkup: childGeneratorLocalName }, importPath);
-    cxt.import({ getReadOnlyProxy: '__getReadOnlyProxy' });
-    const childTagName = node.name;
-
+export function getSlottedContent(
+    node: IrLwcComponent | IrComponent,
+    cxt: TransformerContext
+): EsStatement[] {
     // Anything inside the slotted content is a normal slotted content except for `<template lwc:slot-data>` which is a scoped slot.
     const slottableChildren = node.children.filter((child) => child.type !== 'ScopedSlotFragment');
     const scopedSlottableChildren = node.children.filter(
@@ -94,9 +78,9 @@ export const Component: Transformer<IrComponent> = function Component(node, cxt)
                 draft.attributes = draft.attributes.filter((attr) => attr.name !== 'slot');
             });
             const slotContent = irToEs(clone, cxt);
-            return bAddContent(slotName, null, slotContent);
+            return b.expressionStatement(bAddContent(slotName, null, slotContent));
         } else {
-            return bAddContent(b.literal(''), null, irToEs(child, cxt));
+            return b.expressionStatement(bAddContent(b.literal(''), null, irToEs(child, cxt)));
         }
     });
 
@@ -105,24 +89,16 @@ export const Component: Transformer<IrComponent> = function Component(node, cxt)
         const boundVariable = b.identifier(boundVariableName);
         cxt.pushLocalVars([boundVariableName]);
         // TODO [#4768]: what if the bound variable is `generateMarkup` or some framework-specific identifier?
-        const addContentExpr = bAddContent(
-            child.slotName as EsExpression,
-            boundVariable,
-            irChildrenToEs(child.children, cxt)
+        const addContentExpr = b.expressionStatement(
+            bAddContent(
+                child.slotName as EsExpression,
+                boundVariable,
+                irChildrenToEs(child.children, cxt)
+            )
         );
         cxt.popLocalVars();
         return addContentExpr;
     });
 
-    return [
-        bYieldFromChildGenerator(
-            getChildAttrsOrProps(node.properties, cxt),
-            getChildAttrsOrProps(node.attributes, cxt),
-            shadowSlotContent,
-            lightSlotContent,
-            scopedSlotContent,
-            b.identifier(childGeneratorLocalName),
-            b.literal(childTagName)
-        ),
-    ];
-};
+    return bGenerateSlottedContent(shadowSlotContent, lightSlotContent, scopedSlotContent);
+}
